+++ docs/Changelog.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 4.0.0 TBD
+## 4.1.0 TBD
+- Opportunity to sign Base64 encoded data with `signDataWithDevicePrivateKey`
+  - we added the `dataFormat` parameter with possible `UTF8` and `BASE64` values for the data to be signed
+- `PowerAuthUtils` that provides `getEnvironmentInfo` with device, system and app data
+
+## 4.0.0 (5/2025)
 - Migration to the latest native PowerAuthSDK stack (1.9.x)
   - Use [migration guide](Version-4.0.md) for a smooth migration
-- Possibility to sign Base64 encoded data with `signDataWithDevicePrivateKey`
-  - we added the `dataFormat` parameter with possible `UTF8` and `BASE64` values for the data to be signed
 
 ## 3.1.0 TBD
 - Support for the new React Architecture